Refactor exchange module to integrate core functionality and update price handling
- Introduced `core` module in the `exchange` namespace to encapsulate core exchange logic.
- Updated `ExchangeRuntime` to utilize `ExchangeCore` for order processing and fund management.
- Refactored order handling in tests to accommodate changes in price representation from `f64` to `Price`.
- Modified `MatchingEngine` and `OrderBook` to work with the new `Price` type for better precision and safety.
- Adjusted `OrderManager` to prepare and register orders using the new `Price` type, ensuring consistent price handling.
- Enhanced `Wallet` methods to support operations with the `Price` type, ensuring accurate fund management.
- Removed redundant code and tests related to the old order management and price handling.

+++ src/controllers/exchange_controller.rs
@@ -23,7 +23,7 @@ pub struct BalancedResponse {
 pub struct OrderRequest {
     pub symbol: String,
     pub side: String,
-    pub price: f64,
+    pub price: u64,
     pub quantity: u32,
 }
